requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
jinja2>=3.1.6
python-dotenv>=1.0.1
requests>=2.32.0
pytest>=8.4.2
httpx>=0.28.1
